Move io_serialization to specific files
